+++ modules/core/dependency/python-ihm/test/test_dumper.py
@@ -792,8 +792,8 @@ _ihm_struct_assembly_details.entity_poly_segment_id
         a1 = ihm.AsymUnit(e1)
         system.entities.extend((e1, e2))
         system.asym_units.append(a1)
-        # Note that no asym unit uses entity e2, so the assembly
-        # should omit the chain ID ('.')
+        # Note that no asym unit uses entity e2, so it won't be included
+        # in the assembly
 
         # Assign entity and asym IDs
         ihm.dumper._EntityDumper().finalize(system)
@@ -814,7 +814,6 @@ _ihm_entity_poly_segment.seq_id_end
 _ihm_entity_poly_segment.comp_id_begin
 _ihm_entity_poly_segment.comp_id_end
 1 1 1 3 ALA GLY
-2 2 1 2 GLU TRP
 #
 """)
 
@@ -838,7 +837,6 @@ _ihm_struct_assembly_details.entity_id
 _ihm_struct_assembly_details.asym_id
 _ihm_struct_assembly_details.entity_poly_segment_id
 1 1 1 foo 1 A 1
-2 1 1 bar 2 . 2
 #
 """)
 
@@ -2011,6 +2009,8 @@ _ihm_localization_density_files.entity_poly_segment_id
         a1._id = 'X'
         system.entities.extend((e1, e2, e3))
         system.asym_units.append(a1)
+        system.orphan_features.append(ihm.restraint.ResidueFeature([e2]))
+        system.orphan_features.append(ihm.restraint.NonPolyFeature([e3]))
 
         system._make_complete_assembly()
 
@@ -2019,10 +2019,11 @@ _ihm_localization_density_files.entity_poly_segment_id
         dumper = ihm.dumper._EntityPolySegmentDumper()
         dumper.finalize(system) # assign IDs
 
-        # e1 isn't directly used in the assembly (a1 is used instead) so
-        # should have no range ID
+        # e1 isn't directly used in anything (a1 is used instead, in the
+        # assembly) so should have no range ID
         self.assertFalse(hasattr(e1, '_range_id'))
         self.assertEqual(a1._range_id, 1)
+        # e2 is use, in a ResidueFeature, so should have a range ID
         self.assertEqual(e2._range_id, 2)
         # non-polymers don't have ranges
         self.assertEqual(e3._range_id, None)
@@ -2575,17 +2576,19 @@ _ihm_geometric_object_plane.transformation_id
         a3 = ihm.AsymUnit(e2, 'heme')
         system.asym_units.extend((a1, a2, a3))
 
-        f = ihm.restraint.ResidueFeature([a1, a2(2,3)])
+        f = ihm.restraint.ResidueFeature([a1, a2(2,3), e1, e1(2,3)])
         system.orphan_features.append(f)
         # Cannot make a ResidueFeature that includes a non-polymer 'residue'
         self.assertRaises(ValueError, ihm.restraint.ResidueFeature, [a1, a3])
 
         # Polymeric atom feature
         f = ihm.restraint.AtomFeature([a1.residue(1).atom('CA'),
-                                       a2.residue(2).atom('N')])
+                                       a2.residue(2).atom('N'),
+                                       e1.residue(1).atom('CB')])
         system.orphan_features.append(f)
         # Nonpolymeric atom feature
-        f = ihm.restraint.AtomFeature([a3.residue(1).atom('FE')])
+        f = ihm.restraint.AtomFeature([a3.residue(1).atom('FE'),
+                                       e2.residue(1).atom('FE')])
         system.orphan_features.append(f)
         # Cannot make one feature that selects both polymer and nonpolymer
         self.assertRaises(ValueError, ihm.restraint.AtomFeature,
@@ -2593,7 +2596,7 @@ _ihm_geometric_object_plane.transformation_id
                                        a2.residue(2).atom('N'),
                                        a3.residue(1).atom('FE')])
         # Nonpolymeric feature
-        f = ihm.restraint.NonPolyFeature([a3])
+        f = ihm.restraint.NonPolyFeature([a3, e2])
         system.orphan_features.append(f)
         # Cannot make a NonPolyFeature that includes a polymer 'residue'
         self.assertRaises(ValueError, ihm.restraint.NonPolyFeature, [a1, a3])
@@ -2635,6 +2638,8 @@ _ihm_poly_residue_feature.seq_id_end
 _ihm_poly_residue_feature.comp_id_end
 1 1 1 A 1 ALA 4 THR
 2 1 1 B 2 CYS 3 GLY
+3 1 1 . 1 ALA 4 THR
+4 1 1 . 2 CYS 3 GLY
 #
 #
 loop_
@@ -2647,6 +2652,7 @@ _ihm_poly_atom_feature.comp_id
 _ihm_poly_atom_feature.atom_id
 1 2 1 A 1 ALA CA
 2 2 1 B 2 CYS N
+3 2 1 . 1 ALA CB
 #
 #
 loop_
@@ -2657,7 +2663,9 @@ _ihm_non_poly_feature.asym_id
 _ihm_non_poly_feature.comp_id
 _ihm_non_poly_feature.atom_id
 1 3 2 C HEM FE
-2 4 2 C HEM .
+2 3 2 . HEM FE
+3 4 2 C HEM .
+4 4 2 . HEM .
 #
 #
 loop_