+++ modules/core/dependency/python-ihm/test/test_examples.py
@@ -80,7 +80,7 @@ class Tests(unittest.TestCase):
         # can read it
         with open(out) as fh:
             contents = fh.readlines()
-        self.assertEqual(len(contents), 87)
+        self.assertEqual(len(contents), 64)
         with open(out) as fh:
             s, = ihm.reader.read(fh)
         os.unlink(out)